add task feature backend and frontend
In task site basic task creation with titel, description (optional) and due date (optional) is possible.
Tasks are shown below and can be completed or deleted. A completed task shows undo button in case it was accidentally completed.
Currently no date check is in place for due date - 14.2.1028 would be valid due date.